LPD-92419 Extract RecordingMockHttp to be reused by test classes
